Expansion of Conversion method; genericification
Code unlintered, not formatted

Conversion method now has generic datatype holding conversion information.
Conversion methods are meant to be lambdas, and converts to and from the 'main' datatype.
It has preferred target values.
The intent is to expand the conversion functions - hence currently dead code 'conversion_set_number'
Should be effectivly feature-neutral to users in discord, but not feature-neutral to dev.